Move rounding modes to new module
